app/pyproject.toml
[project]
name = "myfastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.4",
    "fastapi[standard]>=0.115.13",
    "ipdb>=0.13.13",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
]

[tool.uv.workspace]
members = [
    "app",
]
